perf(swish): reduce memory efficiency overhead
We already have parts of the results. Doing two more simple operations to instantly have the gradient makes it much simpler.\
Additionally, in the initial PR it was stated that this implementation is incompatible with JIT. This is not true. Use `torch.jit.trace` instead of `torch.jit.script` to achieve the same effect, and trace over autograd functions (while keeping their memory efficiency).